+++ integration/test_subscription.py
@@ -1,7 +1,3 @@
-# TODO I wrote this before I started putting all of the database queries into
-# a separate module. This could be modified into a unit test and probably
-# grouped into the test_server.py module.
-
 from datetime import datetime, timezone
 import gzip
 import logging